# crates/observation-tools-client/openapi.json
{
  "components": {
    "schemas": {
      "CreateExecutionRequest": {
        "description": "Request to create a new execution",
        "properties": {
          "execution": {
            "$ref": "#/components/schemas/Execution",
            "description": "The execution to create"
          }
        },
        "required": ["execution"],
        "type": "object"
      },
      "CreateExecutionResponse": {
        "description": "Response after creating an execution",
        "type": "object"
      },
      "CreateObservationsRequest": {
        "description": "Request to create observations (can be a batch)",
        "properties": {
          "observations": {
            "description": "List of observations to create",
            "items": {
              "$ref": "#/components/schemas/Observation"
            },
            "type": "array"
          }
        },
        "required": ["observations"],
        "type": "object"
      },
      "CreateObservationsResponse": {
        "description": "Response after creating observations",
        "type": "object"
      },
      "ErrorResponse": {
        "description": "API error response",
        "properties": {
          "code": {
            "description": "Optional error code",
            "nullable": true,
            "type": "string"
          },
          "details": {
            "description": "Optional additional details"
          },
          "error": {
            "description": "Error message",
            "type": "string"
          }
        },
        "required": ["error"],
        "type": "object"
      },
      "Execution": {
        "description": "An execution represents the root scope for data collection.\nAll observations are associated with one execution.",
        "properties": {
          "created_at": {
            "description": "When this execution was created",
            "format": "date-time",
            "type": "string"
          },
          "id": {
            "$ref": "#/components/schemas/ExecutionId",
            "description": "Unique identifier for this execution"
          },
          "metadata": {
            "additionalProperties": {
              "type": "string"
            },
            "description": "User-defined metadata as key-value pairs",
            "propertyNames": {
              "type": "string"
            },
            "type": "object"
          },
          "name": {
            "description": "User-defined name for this execution",
            "type": "string"
          },
          "updated_at": {
            "description": "When this execution was last updated",
            "format": "date-time",
            "type": "string"
          }
        },
        "required": ["id", "name", "created_at", "updated_at"],
        "type": "object"
      },
      "ExecutionId": {
        "description": "Unique identifier for an execution (UUIDv7)",
        "example": "018e9a3a2c1b7e3f8d2a4b5c6d7e8f9a",
        "type": "string"
      },
      "GetExecutionResponse": {
        "description": "Response for getting a single execution",
        "properties": {
          "execution": {
            "$ref": "#/components/schemas/Execution",
            "description": "The execution"
          }
        },
        "required": ["execution"],
        "type": "object"
      },
      "GetObservationResponse": {
        "description": "Response for getting a single observation",
        "properties": {
          "observation": {
            "$ref": "#/components/schemas/Observation",
            "description": "The observation"
          }
        },
        "required": ["observation"],
        "type": "object"
      },
      "ListExecutionsQuery": {
        "description": "Query parameters for listing executions",
        "properties": {
          "limit": {
            "description": "Maximum number of results to return",
            "minimum": 0,
            "nullable": true,
            "type": "integer"
          },
          "offset": {
            "description": "Number of results to skip (for pagination)",
            "minimum": 0,
            "nullable": true,
            "type": "integer"
          },
          "search": {
            "description": "Search query for execution names or metadata",
            "nullable": true,
            "type": "string"
          }
        },
        "type": "object"
      },
      "ListExecutionsResponse": {
        "description": "Response for listing executions",
        "properties": {
          "executions": {
            "description": "List of executions",
            "items": {
              "$ref": "#/components/schemas/Execution"
            },
            "type": "array"
          },
          "has_next_page": {
            "description": "Whether there are more results available",
            "type": "boolean"
          }
        },
        "required": ["executions", "has_next_page"],
        "type": "object"
      },
      "ListObservationsQuery": {
        "description": "Query parameters for listing observations",
        "properties": {
          "labels": {
            "description": "Filter by labels (comma-separated)",
            "nullable": true,
            "type": "string"
          },
          "limit": {
            "description": "Maximum number of results to return",
            "minimum": 0,
            "nullable": true,
            "type": "integer"
          },
          "offset": {
            "description": "Number of results to skip (for pagination)",
            "minimum": 0,
            "nullable": true,
            "type": "integer"
          },
          "parent_span_id": {
            "description": "Filter by parent span ID",
            "nullable": true,
            "type": "string"
          },
          "search": {
            "description": "Search query for observation names or metadata",
            "nullable": true,
            "type": "string"
          },
          "source_file": {
            "description": "Filter by source file",
            "nullable": true,
            "type": "string"
          }
        },
        "type": "object"
      },
      "ListObservationsResponse": {
        "description": "Response for listing observations",
        "properties": {
          "has_next_page": {
            "description": "Whether there are more results available",
            "type": "boolean"
          },
          "observations": {
            "description": "List of observations",
            "items": {
              "$ref": "#/components/schemas/Observation"
            },
            "type": "array"
          }
        },
        "required": ["observations", "has_next_page"],
        "type": "object"
      },
      "Observation": {
        "description": "An observation is a single piece of collected data",
        "properties": {
          "created_at": {
            "description": "When this observation was created",
            "format": "date-time",
            "type": "string"
          },
          "execution_id": {
            "$ref": "#/components/schemas/ExecutionId",
            "description": "ID of the execution this observation belongs to"
          },
          "id": {
            "$ref": "#/components/schemas/ObservationId",
            "description": "Unique identifier for this observation"
          },
          "labels": {
            "description": "Hierarchical labels for grouping observations\nUses path convention (e.g., \"api/request/headers\")",
            "items": {
              "type": "string"
            },
            "type": "array"
          },
          "metadata": {
            "additionalProperties": {
              "type": "string"
            },
            "description": "User-defined metadata as key-value pairs",
            "propertyNames": {
              "type": "string"
            },
            "type": "object"
          },
          "name": {
            "description": "User-defined name for this observation",
            "type": "string"
          },
          "parent_span_id": {
            "description": "Parent span ID (for tracing integration)",
            "nullable": true,
            "type": "string"
          },
          "payload": {
            "$ref": "#/components/schemas/Payload",
            "description": "The observation payload"
          },
          "source": {
            "$ref": "#/components/schemas/SourceInfo",
            "description": "Source location where this observation was created",
            "nullable": true
          }
        },
        "required": ["id", "execution_id", "name", "payload", "created_at"],
        "type": "object"
      },
      "ObservationId": {
        "description": "Unique identifier for an observation (UUIDv7)",
        "example": "018e9a3a2c1b7e3f8d2a4b5c6d7e8f9b",
        "type": "string"
      },
      "Payload": {
        "description": "Payload data for an observation",
        "properties": {
          "data": {
            "description": "The actual payload data",
            "type": "string"
          },
          "mime_type": {
            "description": "MIME type of the payload (e.g., \"text/plain\", \"application/json\")",
            "type": "string"
          },
          "size": {
            "description": "Size of the payload in bytes",
            "minimum": 0,
            "type": "integer"
          }
        },
        "required": ["data", "mime_type", "size"],
        "type": "object"
      },
      "SourceInfo": {
        "description": "Source location information for an observation",
        "properties": {
          "column": {
            "description": "Optional column number",
            "format": "int32",
            "minimum": 0,
            "nullable": true,
            "type": "integer"
          },
          "file": {
            "description": "Source file path",
            "type": "string"
          },
          "line": {
            "description": "Line number in the source file",
            "format": "int32",
            "minimum": 0,
            "type": "integer"
          }
        },
        "required": ["file", "line"],
        "type": "object"
      }
    }
  },
  "info": {
    "description": "API for the Observation Tools developer data inspection toolkit",
    "license": {
      "identifier": "AGPL-3",
      "name": "AGPL-3"
    },
    "title": "Observation Tools API",
    "version": "0.1.0"
  },
  "openapi": "3.0.3",
  "paths": {
    "/api/exe": {
      "get": {
        "operationId": "list_executions",
        "parameters": [
          {
            "description": "Maximum number of results to return",
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "minimum": 0,
              "nullable": true,
              "type": "integer"
            }
          },
          {
            "description": "Number of results to skip (for pagination)",
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "minimum": 0,
              "nullable": true,
              "type": "integer"
            }
          },
          {
            "description": "Search query for execution names or metadata",
            "in": "query",
            "name": "search",
            "required": false,
            "schema": {
              "nullable": true,
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListExecutionsResponse"
                }
              }
            },
            "description": "List of executions"
          },
          "400": {
            "description": "Bad request"
          }
        },
        "summary": "List all executions",
        "tags": ["executions"]
      },
      "post": {
        "operationId": "create_execution",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateExecutionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateExecutionResponse"
                }
              }
            },
            "description": "Execution created successfully"
          },
          "400": {
            "description": "Bad request"
          }
        },
        "summary": "Create a new execution",
        "tags": ["executions"]
      }
    },
    "/api/exe/{execution_id}/obs": {
      "get": {
        "operationId": "list_observations",
        "parameters": [
          {
            "description": "Execution ID",
            "in": "path",
            "name": "execution_id",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "Maximum number of results to return",
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "minimum": 0,
              "nullable": true,
              "type": "integer"
            }
          },
          {
            "description": "Number of results to skip (for pagination)",
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "minimum": 0,
              "nullable": true,
              "type": "integer"
            }
          },
          {
            "description": "Search query for observation names or metadata",
            "in": "query",
            "name": "search",
            "required": false,
            "schema": {
              "nullable": true,
              "type": "string"
            }
          },
          {
            "description": "Filter by labels (comma-separated)",
            "in": "query",
            "name": "labels",
            "required": false,
            "schema": {
              "nullable": true,
              "type": "string"
            }
          },
          {
            "description": "Filter by source file",
            "in": "query",
            "name": "source_file",
            "required": false,
            "schema": {
              "nullable": true,
              "type": "string"
            }
          },
          {
            "description": "Filter by parent span ID",
            "in": "query",
            "name": "parent_span_id",
            "required": false,
            "schema": {
              "nullable": true,
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListObservationsResponse"
                }
              }
            },
            "description": "List of observations"
          },
          "400": {
            "description": "Bad request"
          }
        },
        "summary": "List observations for an execution",
        "tags": ["observations"]
      },
      "post": {
        "operationId": "create_observations",
        "parameters": [
          {
            "description": "Execution ID",
            "in": "path",
            "name": "execution_id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateObservationsRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateObservationsResponse"
                }
              }
            },
            "description": "Observations created successfully"
          },
          "400": {
            "description": "Bad request"
          }
        },
        "summary": "Create observations (batch)",
        "tags": ["observations"]
      }
    },
    "/api/exe/{execution_id}/obs/{observation_id}": {
      "get": {
        "operationId": "get_observation",
        "parameters": [
          {
            "description": "Execution ID",
            "in": "path",
            "name": "execution_id",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "Observation ID",
            "in": "path",
            "name": "observation_id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetObservationResponse"
                }
              }
            },
            "description": "Observation details"
          },
          "400": {
            "description": "Bad request"
          },
          "404": {
            "description": "Observation not found"
          }
        },
        "summary": "Get a single observation",
        "tags": ["observations"]
      }
    },
    "/api/exe/{execution_id}/obs/{observation_id}/content": {
      "get": {
        "operationId": "get_observation_blob",
        "parameters": [
          {
            "description": "Execution ID",
            "in": "path",
            "name": "execution_id",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "Observation ID",
            "in": "path",
            "name": "observation_id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/octet-stream": {
                "schema": {
                  "items": {
                    "format": "int32",
                    "minimum": 0,
                    "type": "integer"
                  },
                  "type": "array"
                }
              }
            },
            "description": "Observation blob content"
          },
          "400": {
            "description": "Bad request"
          },
          "404": {
            "description": "Observation blob not found"
          }
        },
        "summary": "Get observation blob content",
        "tags": ["observations"]
      }
    },
    "/api/exe/{id}": {
      "get": {
        "operationId": "get_execution",
        "parameters": [
          {
            "description": "Execution ID",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetExecutionResponse"
                }
              }
            },
            "description": "Execution details"
          },
          "400": {
            "description": "Bad request"
          },
          "404": {
            "description": "Execution not found"
          }
        },
        "summary": "Get a single execution",
        "tags": ["executions"]
      }
    }
  },
  "tags": [
    {
      "description": "Execution management endpoints",
      "name": "executions"
    },
    {
      "description": "Observation management endpoints",
      "name": "observations"
    }
  ]
}